remove unnecessary extraction
