--- packages/jest-esm-utils/lib/index.js
@@ -1,10 +1,0 @@
-//
-// SPDX-FileCopyrightText: 2025 SAP SE or an SAP affiliate company and Gardener contributors
-//
-// SPDX-License-Identifier: Apache-2.0
-//
-
-export {
-  spyOnHelper,
-  moduleResolve,
-} from './spyOnHelper.js'
