--- packages/jest-esm-utils/package.json
@@ -1,53 +1,0 @@
-{
-  "name": "@gardener-dashboard/jest-esm-utils",
-  "version": "1.0.0",
-  "description": "Utilities for testing",
-  "license": "Apache-2.0",
-  "type": "module",
-  "exports": "./lib/index.js",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/gardener/dashboard.git",
-    "directory": "packages/jest-esm-utils"
-  },
-  "bugs": {
-    "url": "https://github.com/gardener/dashboard/issues"
-  },
-  "publishConfig": {
-    "registry": "https://npm.pkg.github.com/"
-  },
-  "scripts": {
-    "lint": "yarn workspace @gardener-dashboard/packages eslint -c eslint.config.js ./jest-esm-utils --max-warnings=0",
-    "test": "node --experimental-vm-modules $(yarn bin jest)",
-    "test-coverage": "yarn test --coverage"
-  },
-  "devDependencies": {
-    "@jest/globals": "^29.7.0",
-    "jest": "^29.7.0"
-  },
-  "jest": {
-    "restoreMocks": true,
-    "verbose": true,
-    "coverageDirectory": "coverage",
-    "collectCoverageFrom": [
-      "lib/**/*.js"
-    ],
-    "testEnvironment": "node",
-    "transformIgnorePatterns": [
-      "/node_modules/",
-      "\\.pnp\\.[^\\/]+$"
-    ],
-    "coverageThreshold": {
-      "global": {
-        "branches": 84.5,
-        "functions": 97,
-        "lines": 95,
-        "statements": 95
-      }
-    }
-  },
-  "packageManager": "yarn@4.8.1",
-  "engines": {
-    "node": "^22.7.0"
-  }
-}
